requirements.txt
anthropic>=0.40.0
elevenlabs>=1.50.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
sounddevice>=0.5.0
soundfile>=0.12.1
numpy>=1.26.0
fastapi>=0.115.0
uvicorn>=0.30.0
